feat: impl openai async generator
