+++ wren-ai-service/src/providers/llm/openai.py
@@ -1,16 +1,20 @@
 import logging
 import os
-from typing import Any, Dict, List, Optional
+from typing import Any, Callable, Dict, List, Optional, Union
 
 import backoff
 import openai
 from haystack import component
 from haystack.components.embedders import OpenAIDocumentEmbedder, OpenAITextEmbedder
-from haystack.utils.auth import Secret
-from openai import OpenAI
+from haystack.components.generators import OpenAIGenerator
+from haystack.dataclasses import ChatMessage, StreamingChunk
+from haystack.utils import (
+    Secret,
+)
+from openai import AsyncOpenAI, OpenAI, Stream
+from openai.types.chat import ChatCompletion, ChatCompletionChunk
 
 from src.core.provider import LLMProvider
-from src.pipelines.experiment.async_generator import OpenAIGenerator
 from src.providers.loader import provider
 
 logger = logging.getLogger("wren-ai-service")
@@ -27,16 +31,89 @@ EMBEDDING_MODEL_DIMENSION = 3072
 
 
 @component
-class CustomOpenAIGenerator(OpenAIGenerator):
+class AsyncGenerator(OpenAIGenerator):
+    def __init__(
+        self,
+        api_key: Secret = Secret.from_env_var("OPENAI_API_KEY"),
+        model: str = "gpt-3.5-turbo",
+        streaming_callback: Optional[Callable[[StreamingChunk], None]] = None,
+        api_base_url: Optional[str] = None,
+        organization: Optional[str] = None,
+        system_prompt: Optional[str] = None,
+        generation_kwargs: Optional[Dict[str, Any]] = None,
+    ):
+        super(AsyncGenerator, self).__init__(
+            api_key,
+            model,
+            streaming_callback,
+            api_base_url,
+            organization,
+            system_prompt,
+            generation_kwargs,
+        )
+        self.client = AsyncOpenAI(
+            api_key=api_key.resolve_value(),
+            organization=organization,
+            base_url=api_base_url,
+        )
+
     @component.output_types(replies=List[str], meta=List[Dict[str, Any]])
     @backoff.on_exception(backoff.expo, openai.RateLimitError, max_time=60, max_tries=3)
     async def run(
         self, prompt: str, generation_kwargs: Optional[Dict[str, Any]] = None
     ):
         logger.debug(f"Running OpenAI generator with prompt: {prompt}")
-        return await super(CustomOpenAIGenerator, self).run(
-            prompt=prompt, generation_kwargs=generation_kwargs
+        message = ChatMessage.from_user(prompt)
+        if self.system_prompt:
+            messages = [ChatMessage.from_system(self.system_prompt), message]
+        else:
+            messages = [message]
+
+        # update generation kwargs by merging with the generation kwargs passed to the run method
+        generation_kwargs = {**self.generation_kwargs, **(generation_kwargs or {})}
+
+        # adapt ChatMessage(s) to the format expected by the OpenAI API
+        openai_formatted_messages = [message.to_openai_format() for message in messages]
+
+        completion: Union[
+            Stream[ChatCompletionChunk], ChatCompletion
+        ] = await self.client.chat.completions.create(
+            model=self.model,
+            messages=openai_formatted_messages,  # type: ignore
+            stream=self.streaming_callback is not None,
+            **generation_kwargs,
         )
+
+        completions: List[ChatMessage] = []
+        if isinstance(completion, Stream):
+            num_responses = generation_kwargs.pop("n", 1)
+            if num_responses > 1:
+                raise ValueError("Cannot stream multiple responses, please set n=1.")
+            chunks: List[StreamingChunk] = []
+            chunk = None
+
+            # pylint: disable=not-an-iterable
+            for chunk in completion:
+                if chunk.choices and self.streaming_callback:
+                    chunk_delta: StreamingChunk = self._build_chunk(chunk)
+                    chunks.append(chunk_delta)
+                    self.streaming_callback(
+                        chunk_delta
+                    )  # invoke callback with the chunk_delta
+            completions = [self._connect_chunks(chunk, chunks)]
+        elif isinstance(completion, ChatCompletion):
+            completions = [
+                self._build_message(completion, choice) for choice in completion.choices
+            ]
+
+        # before returning, do post-processing of the completions
+        for response in completions:
+            self._check_finish_reason(response)
+
+        return {
+            "replies": [message.content for message in completions],
+            "meta": [message.meta for message in completions],
+        }
 
 
 @provider("openai")
@@ -63,7 +140,7 @@ class OpenAILLMProvider(LLMProvider):
         model_kwargs: Optional[Dict[str, Any]] = GENERATION_MODEL_KWARGS,
         system_prompt: Optional[str] = None,
     ):
-        return CustomOpenAIGenerator(
+        return AsyncGenerator(
             api_key=self._api_key,
             model=self._generation_model,
             system_prompt=system_prompt,
